+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0 - 2019-##-
 
 1. Add `ZERO_DATE` constant. This is used to check for zero dates returned by go servers.
+2. Add `SERVER_URLS_BY_MODE` and `DEV_SERVER_URLS` to define default server URLs for sidebar content.
 
 ## 1.0.0 - 2019-06-17